K8SPS-18: Configure semi-sync nodes through orchestrator
Orchestrator 3.2.6 has a new feature to enforce exact number of
semi-sync nodes. According to `rpl_semi_sync_master_wait_for_slave_count`
variable it enable/disable semi-sync replication on replica nodes.

If the user configures `sizeSemiSync` in `cr.yaml`, the operator enables semi-sync
replication on the primary (`set global rpl_semi_sync_master_enabled=1`)
and set `rpl_semi_sync_master_wait_for_slave_count` to the user defined
value. Afterwards, Orchestrator will detect a failure due to
`EnforceExactSemiSyncReplicas` config option of itself and will try to
configure replicas as semi-sync to reach to the desired state.

If the user sets `sizeSemiSync` to 0 or removes the field, the operator
disables the semi-sync on primary (`set global rpl_semi_sync_master_enabled=0`).
Orchestrator won't disable semi-sync on replicas in this scenario but
it's fine since the replication is async if only one side is semi-sync.

Note that, we're not using the new plugins that are introduced in
8.0.26. Orchestrator doesn't support this plugins yet and until they're
supported by it we need to stay on 8.0.25.