K8SPS-13: Make orchestrator discover MySQL pods
Before this commit we were using peer-list utility to discover new MySQL
pods. Although it was working, it wasn't really reliable and crashed
often.

With these changes, we discover pods by sending requests to Orchestrator
from the operator itself. Unfortunately, the assumption that Orchestrator
will discover every instance if we give it only one host is wrong. So,
we need to send a request for every MySQL pod.