K8SPS-28: Allow custom config for MySQL
We support three different ways to customize MySQL configuration:

1. `spec.mysql.configuration` field in `cr.yaml`. In this case, the operator will
   create the ConfigMap.
2. Creating a ConfigMap named `<cluster-name>-mysql` that contains the key `my.cnf`
3. Creating a Secret named `<cluster-name>-mysql` that contains the key `my.cnf`

We project ConfigMap and Secret in the same volume. In the entrypoint we
concat these files into MySQL config file (`/etc/my.cnf.d/node.cnf`). By
projecting these two resources, users will be able to split their
customizations. Sensitive values can live in the Secrets while the rest
live in the ConfigMap.

To automatically restart the MySQL pods on configuration change, we
calculate the md5 hash from the value of the `spec.mysql.configuration`
and store it in an annotation in the MySQL StatefulSet's PodSpec.
**This only applies to configuration in cr.yaml. Users need to restart
the MySQL pods if they create/update the configmap themselves.**