K8SPS-65: Fix updating replication password with semi-sync replicas (#51)
The operator restarts replication if the user changes replication
password. If semi-sync replication enabled, the primary is locked due to
semi-sync wait count and orchestrator performs an automatic failover.
With this commit we're disabling semi-sync replication on primary before
restarting replication.

Another issue was making sure that replication is not running before
executing `CHANGE REPLICATION SOURCE TO ...`. It was sporadic but
frequent.