K8SPS-70: Fix changing CR secret name with new passwords
If the user changes spec.secretsName and the new secret object contains
updated passwords, MySQL probes start to fail and cluster crashes. We
were mounting the secret referenced in spec.secretsName, with this
commit we mount the internal secret to MySQL and Orchestrator pods.

Also, we're now using the dual password feature that introduced in MySQL 8.
We update the user passwords with `RETAIN CURRENT PASSWORD` and discard
the old passwords after the cluster becomes stable after restarts.