K8SPS-55: Group replication
This commit introduces Group Replication (GR) as a new replication type.

We're using mysql-shell to configure and manage GR. mysql-shell is now
included in the operator image and the operator executes mysql-shell
commands in its shell. Currently we depend on mysql-shell on following
operations:

1. Creating InnoDB cluster
2. Adding new nodes to the cluster
3. Rejoining failed nodes to the cluster
4. Removing nodes from cluster
5. Getting cluster status

If cluster type is GR, the operator won't deploy Orchestrator since the
topology is managed by the GR plugin.

On top of MySQL nodes, the operator deploys MySQL Router to handle
read-write and read-only traffic. The operator won't create a Service
to route traffic to primary pod, all routing will be done in MySQL
Router itself. Users need to use the router service with specific port
numbers.

Currently, **we don't support switching from async to GR and vice
versa.**