K8SPS-87: Don't depend on orchestrator for backups
https://jira.percona.com/browse/K8SPS-87

Also fixed error output in `backupHandler`. The problem with the sidecar was that we were trying to read from `xbcloudErr` in a new goroutine, but if we retrun in the main one before `<-xbcloudDone`, it's not possible because of `defer xbcloudErr.Close()`. And the second problem is that we could write the status code 2 times to `ResponseWriter` in separate goroutines what caused `http: superfluous response.WriteHeader call` error message