K8SPS-113: specify memory size as a template parameter in the config (#120)
* K8SPS-113: specify memory size as a template parameter in the configuration

https://jira.percona.com/browse/K8SPS-113

I decided to use a standard template engine with [sprig](https://masterminds.github.io/sprig/) functions. Currently, only the `Memory` variable is available for use in templates. So, to use for example `Memory*3/4`, we should write `{{int64 (mulf .Memory (divf 3 4))}}`. It's not possible to use templates if `memory` is not specified in `mysql.resources`.

* switch to `github.com/flosch/pongo2`

* add template parameter to `auto-config` test

Co-authored-by: Viacheslav Sarzhan <slava.sarzhan@percona.com>
Co-authored-by: Ege Güneş <ege.gunes@percona.com>