K8SPS-176: remove redundant volume mounts for delete backup job
https://jira.percona.com/browse/K8SPS-176

The `delete-backup` finalizer fails if a cluster was deleted before, because of the missing `cluster1-ssl` secret.
This secret has owner references to the cluster object and gets deleted during cluster deletion.

When the cluster is missing, the `delete-backup` finalizer creates a job that executes `xbcloud` with arguments for deletion. So, we don't need this `cluster1-ssl` secret and any other volume mounts which previously contained this job.