K8SPS-179: Bootstrap Group Replication (#230)
* K8SPS-179: Bootstrap Group Replication

This commit moves group replication configuration logic from operator to
pods. From now, every pod will configure itself for group replication
and join to the cluster.

The need for bootstrap is determined by checking different scenarios:

1. If cluster is not bootstrapped yet and pod is mysql-0, we bootstrap.
   Other pods join to this cluster by their bootstrap processes.
2. If there's only 1 seed in `group_replication_group_seeds`, seed is
   the FQDN of the current pod and pod is mysql-0.
3. `/var/lib/mysql/force-bootstrap` exists.

`force-bootstrap` is a special file for users to select which node to
bootstrap from. This is important in case of full cluster crash. If the
cluster loses the primary and majority to select a new one, we call it
full cluster crash. It's detected in the bootstrap process if no primary
found after iterating over peers. If full cluster crash is detected,
another special file `/var/lib/mysql/full-cluster-crash` is created
automatically with latest `GTID_EXECUTED` as its content. All MySQL
pods start sleeping forever if there's a full cluster crash. User needs
to decide which pod has the newest GTID and create the `force-bootstrap`
file in its datadir to bootstrap the cluster from it. After the
bootstrap, other pods can join to this cluster.

Another important aspect of bootstrapping the group replication is
restoring from backups. In case of GR, we delete the PVCs of pods other
than mysql-0 to force cloning the data. If we don't do that, pods
starting from mysql-1 fails to join to the cluster because they have
more transactions than the mysql-0. These transactions are coming from
the time window between the backup and restore. This is easily
reproducible even with clusters with no traffic at all, since MySQL
Router constantly refreshes the metadata tables inside the MySQL server

* don't use deprecated ioutil

* Update users test.

* Backtick symbols in assert file.

* Remove semicolon from assert file line.

* Use assert ass we get in the diff.

* Add missing env vars to limits test assert files.

* Update gitignore.

* Detect full cluster crash

* Fix users test.

* get password from secret in sysbench pod

* add InnoDBClusterBootstrapped condition

* improve bootstrapping

* add readiness probe for MySQL Router

* enable DEBUG logs for development

* grant select privileges to xtrabackup on replication_group_members

* simplify cluster existence check

* delete PVCs and remove bootstrapped condition after restore

* fix build

* Apply suggestions from code review

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

* update controller-gen

* fix users

* add namespaces and more descriptions to full cluster crash message

* update manifests

* check haproxy status in init-deploy

* update manifests

Co-authored-by: Inel Pandzic <inel.pandzic@percona.com>
Co-authored-by: Viacheslav Sarzhan <slava.sarzhan@percona.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>