Fix backups when storage parameters have special characters
Backups could fail if storage credentials or parameters like `destination`, `containerName`, or  `endpointUrl` contained special characters.

The reason for this is `build/run-backup.sh`. In this script, we send a json request, in which we insert environment variables without escaping them. This way, if a variable contained a character like `"` or `\n`  it could break a json format or provide an invalid value to a parameter.