K8SPS-220: fix backups when storage parameters have special characters (#258)
* Fix backups when storage parameters have special characters

Backups could fail if storage credentials or parameters like `destination`, `containerName`, or  `endpointUrl` contained special characters.

The reason for this is `build/run-backup.sh`. In this script, we send a json request, in which we insert environment variables without escaping them. This way, if a variable contained a character like `"` or `\n`  it could break a json format or provide an invalid value to a parameter.

* Update build/run-backup.sh

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

* add support of format `<bucket-name>/<prefix>` for storages

* add special characters to test's minio deployment

* use custom `minio-secret` for `demand-backup` tests

* create `BucketWithPrefix` type

* fix `demand-backup` test

* fix demand-backup test

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Viacheslav Sarzhan <slava.sarzhan@percona.com>
Co-authored-by: Inel Pandzic <inel.pandzic@percona.com>