K8SPS-400: Fix cluster state getting stuck in initializing if PXB fails
Xtrabackup stops the replication SQL thread as it should when you use
`--safe-slave-backup` flag (we use this flag by default). But if, for
some reason (like using an invalid option), xtrabackup command exits
with non-zero code, it doesn't restart SQL thread. This leaves the
replica in non-replicating state. There's a ticket (PXB-3597) to address
this in xtrabackup. Until it's fixed in PXB, we need to handle this case
and restart the SQL thread if xtrabackup fails with any reason.

Why were cluster getting stuck in initializing state if backup fails?

In case of asynchronous clusters, we check if any instance has any
problems while deciding the cluster state. If one instance is not
replicating from primary, Orchestrator reports this as a problem and
operator sets cluster state to initializing. This is why cluster got
stuck: PXB stops the SQL thread which means replica stops replicating
but since the thread is not restarted after failure. If you manually
start the SQL thread, cluster becomes ready in few seconds. So handling
this case ourselves ensures that cluster state won't be stuck in
initializing.

This made me realize another potential problem: since running a backup
will stop replication on the backup source, any backup that runs for
more than few seconds will make cluster go into initializing state until
backup is finished and replication restarted. This non-optimal for user
experience. This is why I decided to leverage downtime feature in
Orchestrator. With these changes, operator will start downtime for
backup source node and finish the downtime when backup is finished
(success or failure). Operator will ignore any problem in the downtimed
instance so they won't affect the cluster state.

Also in this commit:
* Added `backupSource` field to PerconaServerMySQLBackup status.
* There were two states for failed backups: Error and Failed. Removed
  the Error state and kept Failed.
* Some log and error improvements.